fix 3 tests
